- Added ability to upload images with new park submissions - Updated form components and validations to reflect and require an image to submit - Rearranged files into folders for ease of use
