+++ app/assets/stylesheets/indexImages.scss
@@ -76,12 +76,6 @@ body {
   left:-10px;
   width:250.5px;
   height:50px;
-
-
-
-  @media only screen and (max-width: 250px) {
-
-  }
 }
 
 .ribbonShadowBox {
