+++ app/javascript/react/components/IndexComponents/TagsIndexContainer.js
