+++ app/javascript/react/components/ShowComponents/ParkDetailTile.js
@@ -1,13 +1,13 @@
 import React from 'react'
 
 const ParkDetailTile = (props) => {
-  let image = 'https://images2.imgbox.com/7d/15/MhS0zZLD_o.jpg'
   let ratingDisplay = `Rating: ${props.averageRating}/5`
 
   if (props.averageRating === 0) {
     ratingDisplay = "Be the first to review"
   }
 
+let image = props.image ? props.image.url : null
   return (
     <div>
       <div className="card">
