+++ app/serializers/park_serializer.rb
@@ -1,5 +1,5 @@
 class ParkSerializer < ActiveModel::Serializer
-  attributes :id, :name, :description, :reviews, :average_rating, :currentUser, :city, :state, :country
+  attributes :id, :name, :description, :reviews, :average_rating, :currentUser, :city, :state, :country, :park_photo
 
   has_many :reviews, each_serializer: ReviewSerializer
