+++ db/migrate/20200205205311_image_to_new_park_form.rb
@@ -1,0 +1,5 @@
+class ImageToNewParkForm < ActiveRecord::Migration[5.2]
+  def change
+    add_column :parks, :park_photo, :string, null: false, default: 'https://images2.imgbox.com/f2/28/UZcejNZZ_o.jpg'
+  end
+end
