+++ package.json
@@ -6,10 +6,6 @@
     "yarn": ">=0.25.2"
   },
   "dependencies": {
-    "@rails/webpacker": "3.5",
-    "prop-types": "~15.6.0",
-    "react": "~16.8.0",
-    "react-dom": "~16.8.0",
     "@babel/core": "^7.5.5",
     "@babel/plugin-proposal-class-properties": "^7.0.0",
     "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
@@ -20,18 +16,23 @@
     "@babel/preset-env": "^7.5.5",
     "@babel/preset-react": "^7.0.0",
     "@babel/runtime": "^7.6.2",
+    "@rails/webpacker": "3.5",
     "babel-loader": "^8.0.6",
+    "prop-types": "~15.6.0",
+    "react": "~16.8.0",
+    "react-dom": "~16.8.0",
+    "react-dropzone": "^10.2.1",
     "react-router-dom": "5.0.0",
     "redbox-react": "1.6.0"
   },
   "devDependencies": {
-    "webpack-dev-server": "^2.11.1",
-    "webpack-cli": "^2.0.10",
     "babel-jest": "^24.9.0",
+    "enzyme": "~3.10.0",
+    "enzyme-adapter-react-16": "~1.14.0",
     "fetch-mock": "~5.13.1",
     "jest": "^24.9.0",
-    "enzyme": "~3.10.0",
-    "enzyme-adapter-react-16": "~1.14.0"
+    "webpack-cli": "^2.0.10",
+    "webpack-dev-server": "^2.11.1"
   },
   "scripts": {
     "start": "./bin/webpack-dev-server",
